@@ -29,7 +29,7 @@
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@svitejs/changesets-changelog-github-compact": "^1.2.0",
     "@tailwindcss/typography": "^0.5.16",
-    "@tailwindcss/vite": "^4.1.5",
+    "@tailwindcss/vite": "^4.1.7",
     "@types/d3": "^7.4.3",
     "@types/d3-array": "^3.2.1",
     "@types/d3-color": "^3.1.3",
@@ -56,7 +56,7 @@
     "@types/topojson-specification": "^1.0.5",
     "marked": "^15.0.11",
     "mdsvex": "^0.12.3",
-    "posthog-js": "^1.242.1",
+    "posthog-js": "^1.242.2",
     "prettier": "^3.5.3",
     "prettier-plugin-svelte": "^3.4.0",
     "prism-svelte": "^0.5.0",
@@ -65,12 +65,12 @@
     "rehype-slug": "^6.0.0",
     "shapefile": "^0.6.6",
     "solar-calculator": "^0.3.0",
-    "svelte": "5.29.0",
-    "svelte-check": "^4.2.0",
+    "svelte": "5.30.1",
+    "svelte-check": "^4.2.1",
     "svelte-json-tree": "^2.2.0",
     "svelte-ux": "2.0.0-next.2",
-    "svelte2tsx": "^0.7.38",
-    "tailwindcss": "^4.1.5",
+    "svelte2tsx": "^0.7.39",
+    "tailwindcss": "^4.1.7",
     "topojson-client": "^3.1.0",
     "topojson-simplify": "^3.0.3",
     "tslib": "^2.8.1",
@@ -85,8 +85,7 @@
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
     "@layerstack/svelte-actions": "1.0.1-next.2",
-    "@layerstack/svelte-state": "0.1.0-next.3",
-    "@layerstack/svelte-stores": "1.0.2-next.2",
+    "@layerstack/svelte-state": "0.1.0-next.4",
     "@layerstack/tailwind": "2.0.0-next.4",
     "@layerstack/utils": "1.1.0-next.2",
     "d3-array": "^3.2.4",
@@ -110,7 +109,7 @@
     "d3-time": "^3.1.0",
     "date-fns": "^4.1.0",
     "lodash-es": "^4.17.21",
-    "runed": "^0.27.0"
+    "runed": "^0.28.0"
   },
   "peerDependencies": {
     "svelte": "^5.0.0"
