@@ -1,6 +1,7 @@
 <script lang="ts">
   import { onMount } from 'svelte';
   import posthog from 'posthog-js';
+  import { watch } from 'runed';
 
   import { mdiArrowTopRight, mdiDotsVertical, mdiGithub, mdiTwitter } from '@mdi/js';
   import 'prism-themes/themes/prism-vsc-dark-plus.css';
@@ -16,17 +17,17 @@
     settings,
   } from 'svelte-ux';
   import { sortFunc } from '@layerstack/utils';
-  import { lgScreen } from '@layerstack/svelte-stores';
+  import { MediaQueryPresets } from '@layerstack/svelte-state';
 
   import { dev } from '$app/environment';
   import { afterNavigate, goto } from '$app/navigation';
-  import { page } from '$app/stores';
+  import { page } from '$app/state';
 
   import NavMenu from './_NavMenu.svelte';
 
   import './app.css';
 
-  export let data;
+  let { data, children } = $props();
 
   settings({
     components: {
@@ -82,17 +83,20 @@
 
     // Posthog analytics
     if (!dev) {
-      const unsubscribePage = page.subscribe(($page) => {
-        if (currentPath && currentPath !== $page.url.pathname) {
-          // Page navigated away
+      watch(
+        () => page,
+        () => {
+          if (currentPath && currentPath !== page.url.pathname) {
+            // Page navigated away
+            // @ts-expect-error
+            posthog.capture('$pageleave');
+          }
+          // Page entered
+          currentPath = page.url.pathname;
           // @ts-expect-error
-          posthog.capture('$pageleave');
+          posthog.capture('$pageview');
         }
-        // Page entered
-        currentPath = $page.url.pathname;
-        // @ts-expect-error
-        posthog.capture('$pageview');
-      });
+      );
       const handleBeforeUnload = () => {
         // Hard reloads or browser exit
         // @ts-expect-error
@@ -100,15 +104,16 @@
       };
       window.addEventListener('beforeunload', handleBeforeUnload);
       return () => {
-        unsubscribePage();
         window.removeEventListener('beforeunload', handleBeforeUnload);
       };
     }
   });
+
+  const { lgScreen } = new MediaQueryPresets();
 </script>
 
 <svelte:head>
-  {#if $page.url.origin.includes('https')}
+  {#if page.url.origin.includes('https')}
     <script
       defer
       src="https://static.cloudflareinsights.com/beacon.min.js"
@@ -152,7 +157,7 @@
         <ThemeSelect keyboardShortcuts />
       </div>
 
-      {#if $lgScreen}
+      {#if lgScreen.current}
         <Tooltip title="Discord" placement="left" offset={2}>
           <Button
             icon="M20.33 5.06C18.78 4.33 17.12 3.8 15.38 3.5 15.17 3.89 14.92 4.4 14.74 4.82 12.9 4.54 11.07 4.54 9.26 4.82 9.09 4.4 8.83 3.89 8.62 3.5 6.88 3.8 5.21 4.33 3.66 5.06 0.53 9.79-0.32 14.41 0.1 18.96 2.18 20.52 4.19 21.46 6.17 22.08 6.66 21.4 7.1 20.69 7.48 19.93 6.76 19.66 6.07 19.33 5.43 18.94 5.6 18.81 5.77 18.68 5.93 18.54 9.88 20.39 14.17 20.39 18.07 18.54 18.23 18.68 18.4 18.81 18.57 18.94 17.92 19.33 17.24 19.66 16.52 19.94 16.9 20.69 17.33 21.41 17.82 22.08 19.8 21.46 21.82 20.52 23.9 18.96 24.4 13.69 23.05 9.11 20.33 5.06ZM8.01 16.17C6.83 16.17 5.86 15.06 5.86 13.71 5.86 12.36 6.81 11.25 8.01 11.25 9.22 11.25 10.19 12.36 10.17 13.71 10.17 15.06 9.22 16.17 8.01 16.17ZM15.99 16.17C14.8 16.17 13.83 15.06 13.83 13.71 13.83 12.36 14.78 11.25 15.99 11.25 17.19 11.25 18.17 12.36 18.14 13.71 18.14 15.06 17.19 16.17 15.99 16.17Z"
@@ -217,6 +222,6 @@
   </AppBar>
 
   <main class="isolate" bind:this={mainEl}>
-    <slot />
+    {@render children()}
   </main>
 </AppLayout>
