@@ -281,9 +281,9 @@
       placement: 'bottom',
       variant: 'swatches',
       onclick: (e, item) => {
-        selectedKeys.toggleSelected(item.value);
+        selectedKeys.toggle(item.value);
         // TODO: investigate
-        // selectedSeries.toggleSelected(item.value);
+        // selectedSeries.toggle(item.value);
       },
       onpointerenter: (e, item) => (highlightKey.current = item.value),
       onpointerleave: (e) => (highlightKey.current = null),
